backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
dask==2024.9.0
dask-expr==1.1.14
debugpy==1.8.5
decorator==5.1.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
executing==2.1.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.4.2
Levenshtein==0.25.1
locket==1.0.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
platformdirs==4.3.6
promise==2.3
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg==3.2.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
text-unidecode==1.3
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
